+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Colour run status output using config value `display.color.status.[succeeded|failed|running|skipped|terminated]`. Defaults to `false`.
 - Handling SIGTERM in `exec` by setting run state to `Terminated`.
 - `run kill` to kill runs with status `Running`.
+- `--format` options for `[run|job] list`.
 
 ## [0.2.0] - 2025-11-23